Revert "refacor"
This reverts commit 3c6de7a5bd176d1c7e42470e2fabd6a6adda3eea.

+++ src/main.rs
@@ -78,7 +78,7 @@ async fn main() {
                 ui_chess_board.update_assume_logic(&window_parameters);
                 match ui_chess_board.request_move(&window_parameters) {
                     Some(movement_proposal) => {
-                        let _ = board.move_piece(Coordinate::new(movement_proposal.0 .0, movement_proposal.0 .1), Coordinate::new(movement_proposal.1 .0, movement_proposal.1 .1));
+                        let _ = board.move_piece(Coordinate::new(movement_proposal.0 .0 as i32, movement_proposal.0 .1 as i32), Coordinate::new(movement_proposal.1 .0 as i32, movement_proposal.1 .1 as i32));
                         ui_chess_board.update(&board.squares, &texture);
                     }
                     None => (),
